fix compilation with Java 8

+++ processing/src/test/java/io/druid/query/RetryQueryRunnerTest.java
@@ -131,7 +131,7 @@ public class RetryQueryRunnerTest
           @Override
           public Sequence run(Query query, Map context)
           {
-            if (context.get("count") == 0) {
+            if ((int)context.get("count") == 0) {
               ((List) context.get(RetryQueryRunner.missingSegments)).add(
                   new SegmentDescriptor(
                       new Interval(
